Metrics window: make wrapping a simple checkable action.
https://developer.gnome.org/hig/stable/menus.html.de: "Two linked
actions can be combined into a single menu item, by changing the label
when the item is selected. For example, a Play item may change to Pause.
However, only use this type of item when actions are logical opposites
which are obvious to users. Likewise, do not use this technique for
settings - use check boxes or radio buttons instead."