Make sure tools have a sound enabled/disabled chain
Tools are always enabled once, and disabled once. A tool is disabled when the widget is hidden/closed or when the current tool changes. Enabled when the widget is shown and/or setCurrentTool is called

This is important because we should not change Tool.parent() when the tool is activated (typically because some tools subscribe notifications on the parent glyph, before this change duplicate notifications errors would arise).