Do not try to update the selection label if selection is [None].
Before, the code in FontWindow._selectionChanged() checked for selection
to not be None, but didn't check if it actually was [None]. The user
could trigger an uncaught exception if e.g. he modified some glyph, went
back to the font view, clicked on empty space to deselect all glyphs and
then saved.