[defcon] fix TypeError: newGlyph() missing required pos argument
After this commit 3df77f720b7f2bcfe01eaf4112b36b8bcb3f8462, I am
seeing this error every time I try to import an OTF or TTF,
which uses extractor.

The reason is we can't just assign `Font.newGlyph`, which is a class
attribute, and thus an "unbound" method, to an instance as
self.newGlyph.

In order to dynamically "bind" a method to an instance, we can use
`types.MethodType`.

https://docs.python.org/3/library/types.html

Also, the original newGlyph method is restored with a try/finally
block, in case anything goes wrong during the extraction.